Extract inner anonymous class "HttpConnection" as the more explict named class "LocalHttpConnection"
